Create Day-3
For the 9thquestion, we check every possible pair of numbers in the list to see if their sum matches the target. We avoid checking the same pair twice by starting the second loop just after the first number’s position. Whenever we find a pair that adds up to the target, we print it.

For the 10th question, we compare two lists and find the numbers that appear in the first list but not in the second. We do this by checking each number in the first list and adding it to a new list only if it’s missing from the second list. Finally, we print the new list of unique numbers.